Disable automatic keepalive ping when using ASGI transport
